requirements.txt
astroid==3.3.8
attrs==24.3.0
beautifulsoup4==4.12.3
black==24.10.0
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
dill==0.3.9
idna==3.10
isort==5.13.2
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.14.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycountry==24.6.1
pycparser==2.22
pylint==3.3.3
qrcode==8.0
requests==2.32.3
requests-cache==1.2.1
six==1.17.0
soupsieve==2.6
tomlkit==0.13.2
types-beautifulsoup4==4.12.0.20241020
types-cffi==1.16.0.20241221
types-html5lib==1.1.11.20241018
types-pyOpenSSL==24.1.0.20240722
types-PyYAML==6.0.12.20241221
types-redis==4.6.0.20241004
types-requests==2.32.0.20241016
types-setuptools==75.6.0.20241223
types-ujson==5.10.0.20240515
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.3.0
